Vari-tremolo2i (interpolating) import my old one
I wrote this a while ago by directly hacking Steve's plugin to do what I
wanted. Now I will have to massage it now to fit the sweep.lsp library.
Which already has exp & reversing sweeps but I see I was using direct
sounds not a wavatables for this one due to siosc, so more flexibility
in the sweep.lsp backend is going to be required.